+++ PIC/PowerSupply_variable_pcb_V2.X/Controllers/DAC_Controller.h
@@ -37,8 +37,8 @@ void dacInitDacB(bool buf, uint16_t gain);
 void dacShutDownA();
 void dacShutDownB();
 
-void dacSetVoltageA(float voltage);
-void dacSetVoltageB(float voltage);
+void dacSetVoltageA(double voltage);
+void dacSetVoltageB(double voltage);
